JSONのfloatはC#のdoubleに当たるので修正

+++ NatureRemoMonitor/API/Resource/SensorValue.cs
@@ -1,3 +1,3 @@
 namespace NatureRemoMonitor.API.Resource;
 
-public record SensorValue(float Val, DateTime CreatedAt);
+public record SensorValue(double Val, DateTime CreatedAt);
